refactor(blogpost): group interactives by dir
