+++ src/content/blogpost/chaos-game/interactive/chaos-game-random-restriction/chaos-game-random-restriction.module.css
