+++ src/content/blogpost/chaos-game/interactive/chaos-game-rules/chaos-game-rules-demo.module.css
