+++ src/content/blogpost/chaos-game/interactive/chaos-game-square-restrictions/square-restrictions-demo.module.css
