+++ src/content/blogpost/chaos-game/interactive/index.ts
@@ -1,7 +1,7 @@
-export { default as ChaosGameRulesDemo } from "./ChaosGameRulesDemo.svelte";
-export { default as ChaosRestrictionSubsetsDemo } from "./ChaosRestrictionSubsetsDemo.svelte";
-export { default as PolygonDensityMap } from "./PolygonDensityMap.svelte";
-export { default as ChaosGameRandomRestriction } from "./ChaosGameRandomRestriction.svelte";
-export { default as ChaosGameStepsDemo } from "./ChaosGameStepsDemo.svelte";
-export { default as ChaosGameCreator } from "./ChaosGameCreator.svelte";
-export { default as SquareRestrictionsDemo } from "./SquareRestrictionsDemo.svelte";
+export { default as ChaosGameRulesDemo } from "./chaos-game-rules/ChaosGameRulesDemo.svelte";
+export { default as ChaosRestrictionSubsetsDemo } from "./chaos-game-restriction-subsets/ChaosRestrictionSubsetsDemo.svelte";
+export { default as PolygonDensityMap } from "./polygon-density-map/PolygonDensityMap.svelte";
+export { default as ChaosGameRandomRestriction } from "./chaos-game-random-restriction/ChaosGameRandomRestriction.svelte";
+export { default as ChaosGameStepsDemo } from "./chaos-game-steps/ChaosGameStepsDemo.svelte";
+export { default as ChaosGameCreator } from "./chaos-game-creator/ChaosGameCreator.svelte";
+export { default as SquareRestrictionsDemo } from "./chaos-game-square-restrictions/SquareRestrictionsDemo.svelte";
